# [bench] Adding a cpp benchmark to compare performance of nnc with static and symbolic shapes