--- a/benchmarks/cpp/tensorexpr/CMakeLists.txt
+++ b/benchmarks/cpp/tensorexpr/CMakeLists.txt
@@ -9,6 +9,7 @@ add_executable(
   bench_signed_log1p.cpp
   bench_fuser_overhead.cpp
   bench_gemm.cpp
+  bench_kernels.cpp
   bench_parallel.cpp
   bench_prefix_sum.cpp
   bench_reduce.cpp