Stack from [ghstack](https://github.com/ezyang/ghstack):
* #72198
* __->__ #72197
* #72196

Benchmark results:
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
KernelBench/Eager/32                  2007 ns         2007 ns       350630
KernelBench/Eager/64                  2050 ns         2050 ns       335326
KernelBench/Eager/512                 2196 ns         2196 ns       320889
KernelBench/Eager/2048                2853 ns         2853 ns       233259
KernelBench/StaticShapes/32            544 ns          544 ns      1303151
KernelBench/StaticShapes/64            559 ns          559 ns      1271121
KernelBench/StaticShapes/512           586 ns          586 ns      1149529
KernelBench/StaticShapes/2048          820 ns          820 ns       855455
KernelBench/SymbolicShapes/32          704 ns          704 ns      1008014
KernelBench/SymbolicShapes/64          730 ns          730 ns       981898
KernelBench/SymbolicShapes/512         727 ns          727 ns       893909
KernelBench/SymbolicShapes/2048       1065 ns         1064 ns       648401
```

Differential Revision: [D33951742](https://our.internmc.facebook.com/intern/diff/D33951742)