@navahgar 2022-02-04
> where do the results of the benchmark show up ?

Added eager as well. Updated the PR description with the results.